% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{board_user}
\alias{board_user}
\title{Create a user board for publishing models}
\usage{
board_user(path = Sys.getenv("PETROGRAPHER_BOARD_PATH", ""))
}
\arguments{
\item{path}{Optional path override (uses PETROGRAPHER_BOARD_PATH env var or default)}
}
\value{
pins board object
}
\description{
Create a user board for publishing models
}
