% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{extract_class_names_from_coco}
\alias{extract_class_names_from_coco}
\title{Extract class names from COCO JSON}
\usage{
extract_class_names_from_coco(coco_json_path)
}
\description{
Extract class names from COCO JSON
}
\keyword{internal}
