% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_path}
\alias{get_dataset_path}
\title{Get path to pinned dataset}
\usage{
get_dataset_path(dataset_id, board = "local", version = NULL)
}
\arguments{
\item{dataset_id}{Dataset name}

\item{board}{Pins board (or board object)}

\item{version}{Specific version to retrieve (NULL = latest)}
}
\value{
Path to dataset tar.gz file
}
\description{
Returns filesystem path to a pinned dataset tar.gz file.
The tar.gz should be extracted at training time.
}
