% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_datasets}
\alias{list_datasets}
\title{List pinned datasets}
\usage{
list_datasets(board = "local")
}
\arguments{
\item{board}{Pins board (NULL = local board, "local" = local board)}
}
\description{
Lists all pinned datasets on a board.
}
