% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{list_models}
\alias{list_models}
\title{List available models}
\usage{
list_models(board = NULL)
}
\arguments{
\item{board}{Pins board (NULL = public hub, "local" = local training board)}
}
\description{
List available models
}
