% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{list_trained_models}
\alias{list_trained_models}
\title{List all locally trained models}
\usage{
list_trained_models()
}
\value{
Character vector of pinned model names
}
\description{
List all locally trained models
}
