% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pg_board_user}
\alias{pg_board_user}
\title{Local user board}
\usage{
pg_board_user(path = Sys.getenv("PETRO_PINS_PATH", ""))
}
\description{
This board lives in the user's filesystem (e.g. Dropbox). It is writable and
versioned. Set \code{PETRO_PINS_PATH} to override the default path.
}
