% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pg_dataset_catalog}
\alias{pg_dataset_catalog}
\title{Build a catalog tibble for published datasets}
\usage{
pg_dataset_catalog(board = pg_board_pkgdown())
}
\arguments{
\item{board}{Pins board to read from (defaults to pkgdown board).}
}
\value{
Tibble with dataset summary information.
}
\description{
Build a catalog tibble for published datasets
}
