% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pg_model_catalog}
\alias{pg_model_catalog}
\title{Build a catalog tibble for published models}
\usage{
pg_model_catalog(board = pg_board_pkgdown())
}
\arguments{
\item{board}{Pins board to read from (defaults to pkgdown board).}
}
\value{
Tibble compatible with \code{bs4cards::bs4cards()}.
}
\description{
Build a catalog tibble for published models
}
