% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict}
\alias{predict}
\title{Predict objects in an image}
\usage{
predict(model, image_path, ...)
}
\arguments{
\item{model}{A PetrographyModel object}

\item{image_path}{Path to image file}

\item{...}{Additional arguments passed to methods}
}
\value{
Tibble with detection results
}
\description{
S3 generic for running predictions with a PetrographyModel.
}
