% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{publish_model}
\alias{publish_model}
\title{Publish model to board}
\usage{
publish_model(
  model_dir,
  model_id,
  board,
  metadata = list(),
  update_manifest = TRUE
)
}
\arguments{
\item{model_dir}{Directory with model files}

\item{model_id}{Name for the model}

\item{board}{Pins board to publish to}

\item{metadata}{Optional metadata list}

\item{update_manifest}{Update _pins.yaml for board_url access}
}
\description{
Publish model to board
}
