% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_model_hpc}
\alias{train_model_hpc}
\title{Train model on HPC using SLURM}
\usage{
train_model_hpc(
  data_dir,
  dataset_id,
  model_id,
  run_id,
  max_iter,
  learning_rate,
  num_classes,
  backbone,
  freeze_at,
  eval_period,
  checkpoint_period,
  ims_per_batch,
  num_workers,
  hpc_cpus_per_task,
  hpc_mem,
  gpus,
  workspace_dir
)
}
\description{
Train model on HPC using SLURM
}
\keyword{internal}
