% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_model_local}
\alias{train_model_local}
\title{Train model locally}
\usage{
train_model_local(
  data_dir,
  model_id,
  model_variant,
  resolution,
  epochs,
  batch_size,
  grad_accum_steps,
  learning_rate,
  device,
  workspace_dir,
  use_amp,
  amp_dtype,
  gradient_checkpointing,
  num_workers,
  validate_every,
  early_stopping_patience
)
}
\description{
Train model locally
}
\keyword{internal}
