% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{model_info}
\alias{model_info}
\title{Get model info}
\usage{
model_info(model_id, board = NULL)
}
\arguments{
\item{model_id}{Model name}

\item{board}{Pins board (NULL = public hub)}
}
\description{
Get model info
}
