% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{prepare_training_config}
\alias{prepare_training_config}
\title{Train model locally using available hardware}
\usage{
prepare_training_config(
  data_dir,
  dataset_id,
  dataset_version,
  model_id,
  num_classes,
  backbone,
  freeze_at,
  max_iter,
  learning_rate,
  device,
  eval_period,
  checkpoint_period,
  ims_per_batch,
  num_workers,
  hpc_cpus_per_task,
  hpc_mem,
  gpus
)
}
\description{
Train model locally using available hardware
}
\keyword{internal}
