% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{prepare_training_config}
\alias{prepare_training_config}
\title{Train model locally using available hardware}
\usage{
prepare_training_config(
  data_dir,
  dataset_id,
  dataset_version,
  model_id,
  model_variant,
  resolution,
  epochs,
  grad_accum_steps,
  learning_rate,
  device,
  batch_size,
  use_amp,
  amp_dtype,
  gradient_checkpointing,
  num_workers,
  time_hours,
  validate_every,
  early_stopping_patience
)
}
\description{
Train model locally using available hardware
}
\keyword{internal}
