% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{prepare_training_config}
\alias{prepare_training_config}
\title{Train model locally using available hardware}
\usage{
prepare_training_config(
  data_dir,
  output_name,
  num_classes,
  backbone,
  freeze_at,
  max_iter,
  learning_rate,
  device,
  eval_period,
  checkpoint_period,
  ims_per_batch,
  num_workers,
  hpc_env,
  hpc_cpus_per_task,
  hpc_mem,
  gpus,
  hpc_host,
  hpc_user,
  hpc_base_dir,
  local_output_dir,
  auto_version,
  publish_after_train,
  model_board,
  model_description
)
}
\description{
Train model locally using available hardware
}
\keyword{internal}
